refactor(profile/utils): rename `getAfterParam` -> `afterParamFactory` and move to `/factories`
