+++ app/profile/[id]/top/albums/page.tsx
@@ -6,11 +6,13 @@ import { getServerToken } from '@app/auth'
 import { STATS_MEASUREMENT, TIME_RANGE, VIEW } from '@app/profile/constants'
 import { ItemsSection } from '@app/profile/sections'
 import { StatsProvider, type ProfilePageProps } from '@app/profile/types'
-import { getAfterParam } from '@app/profile/utils/get-after-param'
-import { validateStatsMeasurement } from '@app/profile/utils/stats-measurement'
-import { validateTimeRange } from '@app/profile/utils/time-range'
-import { validateView } from '@app/profile/utils/view'
 import { validateId } from '@app/utils/validate-id'
+import {
+  validateStatsMeasurement,
+  validateTimeRange,
+  validateView,
+} from '@app/profile/utils/validators'
+import { afterParamFactory } from '@app/profile/utils/factories'
 
 export const runtime = 'edge'
 
@@ -33,7 +35,7 @@ export default async function ProfileTopAlbumsPage({
   if (!token) redirect('/')
 
   const items = await getRigtchTopAlbums(token, {
-    after: getAfterParam(timeRange as RigtchTimeRange),
+    after: afterParamFactory(timeRange as RigtchTimeRange),
     userId,
     limit: 100,
     measurement: statsMeasurement,
