+++ app/profile/[id]/top/artists/loading.tsx
@@ -4,7 +4,7 @@ import { useSearchParams } from 'next/navigation'
 
 import { TopArtistsSkeletonSection } from '@app/profile/sections'
 import { VIEW } from '@app/profile/constants'
-import { validateView } from '@app/profile/utils/view'
+import { validateView } from '@app/profile/utils/validators'
 
 export default function ProfileTopArtistsLoading() {
   const searchParams = useSearchParams()
