+++ app/profile/utils/factories/index.ts
@@ -1,0 +1,1 @@
+export * from './after-param-factory'
