+++ app/profile/utils/validators/validate-stats-measurement.spec.ts
@@ -1,0 +1,16 @@
+import { validateStatsMeasurement } from './validate-stats-measurement'
+
+import { StatsMeasurement } from '@app/api/types'
+
+describe('validateStatsMeasurement', () => {
+  const statsMeasurement = StatsMeasurement.PLAYS
+  const invalid = 'invalid'
+
+  test('should return statsMeasurement when statsMeasurement is a StatsMeasurement', () => {
+    expect(validateStatsMeasurement(statsMeasurement)).toEqual(statsMeasurement)
+  })
+
+  test('should return StatsMeasurement.PLAYS when statsMeasurement is not a StatsMeasurement', () => {
+    expect(validateStatsMeasurement(invalid)).toEqual(StatsMeasurement.PLAYS)
+  })
+})
